requirements.txt
click==8.1.7
colorama==0.4.6
Flask==2.1.0
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
python-dotenv==0.19.2
SQLAlchemy==1.4.46
typing_extensions==4.12.2
werkzeug==2.0.3
WTForms==3.2.1
email-validator
